% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playTheGame.R
\name{playTheGame}
\alias{playTheGame}
\title{Function to play The Game in R.}
\usage{
playTheGame()
}
\description{
Interactive function to play The Game
using console and plot window.
}
\examples{
playTheGame()
}
\author{
Edwin Graham (edwingraham1984@gmail.com)
}
